refactor(scheduler): files organization
