+++ src/main/kotlin/com/github/sylux6/watanabot/scheduler/QuartzScheduler.kt
@@ -1,5 +1,8 @@
 package com.github.sylux6.watanabot.scheduler
 
+import com.github.sylux6.watanabot.scheduler.jobs.DailyTasks
+import com.github.sylux6.watanabot.scheduler.jobs.MaiaBirthday
+import com.github.sylux6.watanabot.scheduler.jobs.WatanabeYouBirthday
 import org.quartz.CronScheduleBuilder.cronSchedule
 import org.quartz.DateBuilder
 import org.quartz.JobBuilder.newJob
@@ -17,9 +20,9 @@ object QuartzScheduler {
         val maiaBirthday = newTrigger().startAt(startTime).withSchedule(cronSchedule("0 0 0 14 3 ? *")).build()
         val dailyTrigger = newTrigger().startAt(startTime).withSchedule(cronSchedule("0 0 0 * * ?")).build()
 
-        scheduler.scheduleJob(newJob(WatanabeYouBirthdayJob::class.java).build(), youBirthday)
-        scheduler.scheduleJob(newJob(MaiaBirthdayJob::class.java).build(), maiaBirthday)
-        scheduler.scheduleJob(newJob(DailyTask::class.java).build(), dailyTrigger)
+        scheduler.scheduleJob(newJob(WatanabeYouBirthday::class.java).build(), youBirthday)
+        scheduler.scheduleJob(newJob(MaiaBirthday::class.java).build(), maiaBirthday)
+        scheduler.scheduleJob(newJob(DailyTasks::class.java).build(), dailyTrigger)
         scheduler.start()
     }
 }
